Try to associate component to a boundary
Here we try a method for associating a boundary with a
component. This could then be treated like the component bodies
in order to create the Impedance BC.

However, For some reason, CurrentModel % Boundaries is of zero length
at CircuitUtils.F90:205
```fortran
print *, "SIZE(CurrentModel % Boundaries)", SIZE(CurrentModel % Boundaries)
```

Why?